Fixed screen ratio calculation
If screen ratio was below one (vertically stretched)
It would be rounded down to 0 causing divide by zero errors

+++ defines.h
@@ -9,7 +9,7 @@
 #define SCREENXPOS    170
 #define SCREENYPOS    30
 
-#define SCREENRATIO   SCREENWIDTH / SCREENHEIGHT
+#define SCREENRATIO   ((float)SCREENWIDTH / (float)SCREENHEIGHT)
 
 #define MAPWIDTH      170
 #define MAPHEIGHT     170
